fix(pr70): move Trap 135 stack-scope regression into tests/
checks/tests are executed as standalone scripts in CI without the minefield
package on PYTHONPATH. Keep preflight regressions there; put compile_registry /
contract_for_match coverage under tests/.
